Converted CriteriaDocTests to CriteriaDocSpec and PlasticCriteriaTests to PlasticCriteriaSpec (conversion of JUnitTests to Spock Tests)
